ai_health_assistant/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7
langchain-google-genai==2.0.3
faiss-cpu==1.12.0
pydantic==2.9.2
typing-extensions==4.12.2

